# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ticketopia API
  version: 1.5.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Read-only event discovery across all Ticketmaster categories, shared with the
    web UI. Event, artist, and venue IDs are provider-scoped (ticketmaster:SOURCE_ID).
    Category and genre IDs are Ticketmaster
    taxonomy IDs. Data is cached, not a live inventory or durable event history.
    Opt-in Google authentication provides MariaDB-backed profiles and private
    preferences, private durable event saves, independent event interest, and public event recommendations.
    Interest identity is private by default; counts include private choices but
    participant lists and profile activity include only explicitly public choices.
    Account APIs use a browser session or a user-scoped API token;
    discovery remains public. Account APIs return 503 when accounts are disabled.
servers:
  - url: /
security: []
paths:
...
  /api/v1/me/saved-events:
    get:
      operationId: listSavedEvents
      summary: List your private saved events
      description: |
        Owner-only last-known event snapshots, newest saved first with a binary
        event-ID tie-breaker. No provider requests are made. Repeated saves keep
        saved_at unchanged. Cursors are bound to the owner; pass next_cursor to
        continue. New saves appear on refresh; removals can shorten later pages.
        Unknown/repeated parameters are rejected. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 512}}
      responses:
        '200':
          description: Your saved collection, including an empty array when no events are saved.
          headers:
            Cache-Control: {description: Private responses must not be stored., schema: {type: string, const: 'private, no-store'}}
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/SavedEvent'}}
                  next_cursor: {type: [string, 'null']}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '503': {$ref: '#/components/responses/SavesUnavailable'}
  /api/v1/me/saved-events/states:
    get:
      operationId: getSavedEventStates
      summary: Read your saved state for a bounded batch of events
      description: |
        Owner-only state without provider requests or metadata. Only event_id is
        accepted; repeat it for up to 101 entries. Absent keys in states mean not
        saved. Browser responses include the current CSRF token to revalidate
        restored controls; bearer responses omit it. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - name: event_id
          in: query
          style: form
          explode: true
          schema:
            type: array
            maxItems: 101
            items: {$ref: '#/components/schemas/ResourceID'}
      responses:
        '200':
          description: Only the authenticated owner's bookmark state.
          content:
            application/json:
              schema:
                type: object
                required: [states]
                properties:
                  states: {type: object, additionalProperties: {type: boolean}}
                  csrf_token: {type: string}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '503': {$ref: '#/components/responses/SavesUnavailable'}
  /api/v1/me/saved-events/{event_id}:
    parameters:
      - {name: event_id, in: path, required: true, schema: {$ref: '#/components/schemas/ResourceID'}}
    get:
      operationId: getSavedEvent
      summary: Read one of your private saved events
      description: No query parameters. Returns 404 for an event not saved by you, regardless of another account's saves. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      responses:
        '200':
          description: Your bookmark and its last-known event snapshot.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SavedEvent'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/SavedEventNotFound'}
        '503': {$ref: '#/components/responses/SavesUnavailable'}
    put:
      operationId: saveEvent
      summary: Save an event privately
      description: |
        Empty body and no query parameters. Resolve server-owned metadata, then
        atomically store event identity, a full last-known snapshot and an owner-only
        bookmark. Retry-safe: repeated PUT preserves saved_at and needs no provider
        call once saved. A retained snapshot can be saved during an outage.
        Does not publish interest, recommendations or attendance or reserve tickets.
        Cookie writes require X-CSRF-Token and same-origin checks; bearer writes do not.
        All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '201':
          description: Bookmark created.
          headers:
            Location: {description: Owner-only bookmark resource., schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SavedEvent'}
        '200':
          description: Existing bookmark, with its original saved_at.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SavedEvent'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFRejected'}
        '404': {$ref: '#/components/responses/SavedEventNotFound'}
        '503': {$ref: '#/components/responses/SavesUnavailable'}
    delete:
      operationId: removeSavedEvent
      summary: Remove an event from your private collection
      description: Empty body and no query parameters. Idempotent and owner-only; absent bookmarks are successful no-ops. Does not remove durable event metadata. Cookie writes require CSRF and same-origin checks. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Bookmark removed or already absent.}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFRejected'}
        '503': {$ref: '#/components/responses/SavesUnavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ticketopia_session
      description: 'HttpOnly, Secure, SameSite=Lax cookie issued after Google sign-in. Local HTTP development uses ticketopia_session instead.'
    bearerToken:
      type: http
      scheme: bearer
      description: Opaque, expiring user-scoped API credential created in an authenticated browser. Not a Google access token or ID token.
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: Required for every cookie-authenticated write. Read csrf_token from GET /api/v1/me. Not required with an API bearer token. Supplied Origin/Referer must match AUTH_BASE_URL.
      schema: {type: string}
  responses:
...
    SavedEventNotFound:
      description: This event cannot be resolved for saving, or is not saved by the authenticated owner for a bookmark read.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    SavesUnavailable:
      description: Accounts/saves are disabled, durable storage is unavailable, or provider data cannot be resolved and no last-known snapshot exists. No in-memory fallback.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    AccountBadRequest:
      description: Unsupported query, invalid JSON/form fields or values, or a request body exceeding 16 KiB.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthenticated:
      description: Missing, expired, revoked, wrong-kind, or invalid credentials. Credential management requires a browser session.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    CSRFRejected:
      description: Invalid or missing CSRF token, or a cross-origin cookie write.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
...
    SavedEvent:
      type: object
      required: [event, saved_at, meta]
      properties:
        event: {$ref: '#/components/schemas/Event'}
        saved_at: {type: string, format: date-time, description: 'Original bookmark creation time, unchanged by repeated saves.'}
        meta:
          allOf:
            - {$ref: '#/components/schemas/Freshness'}
          description: Snapshot collection time. stale is always true for stored bookmark reads; no current-provider freshness is implied.
...
    ResourceID:
      type: string
      pattern: '^ticketmaster:[A-Za-z0-9_-]{1,128}$'
    Freshness:
      type: object
      required: [data_as_of, stale]
      properties:
        data_as_of: {type: string, format: date-time, description: 'When this data was successfully collected, not when this response was served.'}
        stale: {type: boolean, description: A refresh failed or was paused and retained data is being served.}
    Source:
      type: object
      required: [provider, id, url]
      properties:
        provider: {type: string, const: ticketmaster}
        id: {type: string}
        url: {type: string, description: 'Original source URL, or an empty string if unavailable.'}
    NamedID:
      type: object
      required: [id, name]
      properties:
        id: {type: string}
        name: {type: string}
...
    NullableInstant:
      type: [string, 'null']
      format: date-time
    EventStart:
      type: object
      required: [date_time, local_date, local_time, timezone, date_tba, date_tbd, time_tba, no_specific_time]
      properties:
        date_time: {$ref: '#/components/schemas/NullableInstant'}
        local_date: {type: [string, 'null'], description: Provider local date; TBA/TBD flags override any placeholder value.}
        local_time: {type: [string, 'null'], description: Provider local time; time_tba and no_specific_time override placeholder values.}
        timezone: {type: [string, 'null']}
        date_tba: {type: boolean}
        date_tbd: {type: boolean}
        time_tba: {type: boolean}
        no_specific_time: {type: boolean}
    PublicSale:
      type: object
      required: [start, end, start_tbd]
      properties:
        start: {$ref: '#/components/schemas/NullableInstant'}
        end: {$ref: '#/components/schemas/NullableInstant'}
        start_tbd: {type: boolean}
    Presale:
      type: object
      required: [name, description, url, start, end]
      properties:
        name: {type: string}
        description: {type: string}
        url: {type: string}
        start: {$ref: '#/components/schemas/NullableInstant'}
        end: {$ref: '#/components/schemas/NullableInstant'}
    PriceRange:
      type: object
      required: [type, currency, min, max, fees_included]
      properties:
        type: {type: string}
        currency: {type: string}
        min: {type: [string, 'null'], description: Decimal amount preserved without floating-point conversion.}
        max: {type: [string, 'null'], description: Decimal amount preserved without floating-point conversion.}
        fees_included: {type: [boolean, 'null'], description: Null when the source does not establish fee inclusion.}
    Image:
      type: object
      required: [url, width, height, ratio, fallback]
      properties:
        url: {type: string}
        width: {type: integer}
        height: {type: integer}
        ratio: {type: string}
        fallback: {type: boolean}
    Artist:
      type: object
      required: [id, name, source]
      properties:
        id: {$ref: '#/components/schemas/ResourceID'}
        name: {type: string}
        source: {$ref: '#/components/schemas/Source'}
    Place:
      type: object
      required: [name, address, city, state, country, country_code, postal_code, latitude, longitude]
      properties:
        name: {type: string}
        address: {type: string}
        city: {type: string}
        state: {type: string}
        country: {type: string}
        country_code: {type: string}
        postal_code: {type: string}
        latitude: {type: string}
        longitude: {type: string}
    Venue:
      allOf:
        - {$ref: '#/components/schemas/Place'}
        - type: object
          required: [id, source, timezone]
          properties:
            id: {$ref: '#/components/schemas/ResourceID'}
            source: {$ref: '#/components/schemas/Source'}
            timezone: {type: string}
    Classification:
      type: object
      required: [primary, segment, genre, subgenre]
      properties:
        primary: {type: boolean}
        segment: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
        genre: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
        subgenre: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
    Event:
      type: object
      required: [id, name, source, start, status, public_sale, presales, price_ranges, images, artists, venues, place, classifications, info, please_note]
      properties:
        id: {$ref: '#/components/schemas/ResourceID'}
        name: {type: string}
        source: {$ref: '#/components/schemas/Source'}
        start: {$ref: '#/components/schemas/EventStart'}
        status: {type: string, description: 'Source status (onsale, offsale, canceled, postponed, rescheduled) or unknown. Not seat availability.'}
        public_sale: {$ref: '#/components/schemas/PublicSale'}
        presales: {type: array, items: {$ref: '#/components/schemas/Presale'}}
        price_ranges: {type: [array, 'null'], items: {$ref: '#/components/schemas/PriceRange'}, description: 'Null means no advertised prices, not free admission.'}
        images: {type: array, items: {$ref: '#/components/schemas/Image'}}
        artists: {type: array, items: {$ref: '#/components/schemas/Artist'}, description: 'Provider attractions, including performers or teams. The field name is retained for compatibility; events without attractions return an empty array.'}
        venues: {type: array, items: {$ref: '#/components/schemas/Venue'}}
        place: {anyOf: [{$ref: '#/components/schemas/Place'}, {type: 'null'}]}
        classifications: {type: array, items: {$ref: '#/components/schemas/Classification'}}
        info: {type: string}
        please_note: {type: string}
    Problem:
      type: object
      required: [type, title, status, code, detail]
      properties:
        type: {type: string, const: 'about:blank'}
        title: {type: string}
        status: {type: integer}
        code: {type: string, enum: [invalid_query, event_not_found, provider_unavailable, invalid_input, unauthenticated, csrf_failed, invalid_sign_in, account_not_found, rate_limited, credential_limit, accounts_disabled, accounts_unavailable]}
        detail: {type: string}
        fields: {type: object, additionalProperties: {type: string}}